refactor: move utils to internal
