+++ packages/core/src/helpers.ts
@@ -1,7 +1,10 @@
 export { viewExecutionResults } from "./internal/deployment/utils";
 export { createServices } from "./internal/services/createServices";
-export { serializeReplacer } from "./utils/serialize";
-export { IgnitionError, IgnitionValidationError } from "./utils/errors";
+export { serializeReplacer } from "./internal/utils/serialize";
+export {
+  IgnitionError,
+  IgnitionValidationError,
+} from "./internal/utils/errors";
 export { TransactionsService } from "./internal/services/TransactionsService";
 export { ContractsService } from "./internal/services/ContractsService";
 export { VertexResultEnum } from "./internal/types/graph";
