+++ packages/core/src/internal/execution/utils.ts
@@ -1,8 +1,8 @@
 import hash from "object-hash";
 
 import { getDependenciesFor } from "../../internal/graph/adjacencyList";
-import { serializeReplacer } from "../../utils/serialize";
 import { ExecutionGraph } from "../execution/ExecutionGraph";
+import { serializeReplacer } from "../utils/serialize";
 
 export function allDependenciesCompleted(
   vertexId: number,
