+++ packages/core/src/internal/utils/errors.ts
