+++ packages/core/src/internal/utils/serialize.ts
