+++ packages/core/src/validation/dispatch/validateArtifactContract.ts
@@ -2,12 +2,12 @@ import { ethers, BigNumber } from "ethers";
 
 import { ArtifactContractDeploymentVertex } from "../../internal/types/deploymentGraph";
 import { VertexResultEnum } from "../../internal/types/graph";
+import { isArtifact, isParameter } from "../../internal/utils/guards";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
   ValidationVertexVisitResult,
 } from "../../types/validation";
-import { isArtifact, isParameter } from "../../utils/guards";
 
 import { buildValidationError, validateBytesForArtifact } from "./helpers";
