+++ packages/core/test/deploymentBuilder/buildModule.ts
@@ -3,7 +3,7 @@ import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
-import { IgnitionError } from "../../src/utils/errors";
+import { IgnitionError } from "../../src/internal/utils/errors";
 
 describe("deployment builder - buildModule", () => {
   it("should throw if build module is given an async callback", () => {
