+++ packages/core/test/deploymentBuilder/getBytesForArtifact.ts
@@ -7,7 +7,7 @@ import {
   IDeploymentBuilder,
   IDeploymentGraph,
 } from "../../src/internal/types/deploymentGraph";
-import { isCall, isHardhatContract } from "../../src/utils/guards";
+import { isCall, isHardhatContract } from "../../src/internal/utils/guards";
 
 import { getDeploymentVertexByLabel } from "./helpers";
