+++ packages/core/test/deploymentBuilder/parameters.ts
@@ -8,8 +8,8 @@ import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
-import { IgnitionError } from "../../src/utils/errors";
-import { isCallable } from "../../src/utils/guards";
+import { IgnitionError } from "../../src/internal/utils/errors";
+import { isCallable } from "../../src/internal/utils/guards";
 
 describe("deployment builder - parameters", function () {
   let deploymentGraph: IDeploymentGraph;
