+++ packages/core/test/helpers.ts
@@ -10,9 +10,9 @@ import {
   IArtifactsService,
   IAccountsService,
 } from "../src/internal/types/services";
+import { IgnitionError } from "../src/internal/utils/errors";
 import { Artifact } from "../src/types/hardhat";
 import { HasParamResult } from "../src/types/providers";
-import { IgnitionError } from "../src/utils/errors";
 
 export function getMockServices() {
   const mockServices: Services = {
