+++ packages/core/test/tx-sender.ts
@@ -3,8 +3,8 @@ import { assert } from "chai";
 import { ethers } from "ethers";
 import sinon from "sinon";
 
+import { TxSender } from "../src/internal/utils/tx-sender";
 import { GasProvider } from "../src/types/providers";
-import { TxSender } from "../src/utils/tx-sender";
 
 class SignerSpy {
   private _index = -1;
