+++ packages/core/test/validation.ts
@@ -8,11 +8,11 @@ import sinon from "sinon";
 import { buildModule } from "../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../src/internal/process/generateDeploymentGraphFrom";
 import { Services } from "../src/internal/types/services";
+import { IgnitionValidationError } from "../src/internal/utils/errors";
 import { ArtifactContract } from "../src/types/future";
 import { Artifact } from "../src/types/hardhat";
 import { Module, ModuleDict } from "../src/types/module";
 import { ValidationVisitResult } from "../src/types/validation";
-import { IgnitionValidationError } from "../src/utils/errors";
 import { validateDeploymentGraph } from "../src/validation/validateDeploymentGraph";
 
 import { getMockServices } from "./helpers";
